+++ packages/web/visualizations/VisualizationContainer.tsx
@@ -5,9 +5,6 @@ import { useVisualizationDimensions } from '@cbhq/cds-common/visualizations/useV
 import { useDimensions } from '../hooks/useDimensions';
 import { Box } from '../layout';
 
-/**
- * @deprecated this component will be removed from CDS in v6.0.0. It has been moved to cds-web-sparkline.
- */
 /*
 Some visualizations need a static width to render. This container can be dynamically sized and it will inject its static calculated dimensions into its child
  */